Issue an error if a PIPE char is used with a single string.
The following syntax used to be supported in the first iteration of Beancount:

  2014-06-04 * "Payee" |
    ...

Now, to account for more flexibility in the parser, we don't use the PIPE symbol
anymore. We allow tags and links and strings in any arbitraty order. In fact,
this

  2014-06-04 * "Payee" | "Narration"
    ...

can be replaced by this:

  2014-06-04 * "Payee" "Narration"
    ...

But... it'll happen sometimes that I'm still using the PIPE.
This change now issues a warning if a PIPE has been seen on that line
but only a single string has been provided.