Created temporary 'print_prices' report, in order to allow building a dashboard.
Building a dashboard is a bit of an urgent task; I need it now. Down the road,
the 'print_prices' report will just be an instance of the filter view with
"type:price" as the filter, so an explicit report will not be needed. But I need
to be able to dump the prices database now for the dashboard to load, so there
it is.

--HG--
branch : holdings_io